move dilation to BaseConv..LayerSpace
