+++ arbiter/arbiter-deeplearning4j/src/main/java/org/deeplearning4j/arbiter/layers/BaseConvolutionLayerSpace.java
@@ -36,6 +36,7 @@ import org.deeplearning4j.nn.conf.layers.FeedForwardLayer;
 @EqualsAndHashCode(callSuper = true)
 @NoArgsConstructor(access = AccessLevel.PROTECTED) //For Jackson JSON/YAML deserialization
 public abstract class BaseConvolutionLayerSpace<T extends FeedForwardLayer> extends FeedForwardLayerSpace<T> {
+    protected ParameterSpace<int[]> dilation;
     protected ParameterSpace<int[]> kernelSize;
     protected ParameterSpace<int[]> stride;
     protected ParameterSpace<int[]> padding;
@@ -44,6 +45,7 @@ public abstract class BaseConvolutionLayerSpace<T extends FeedForwardLayer> exte
 
     protected BaseConvolutionLayerSpace(Builder builder) {
         super(builder);
+        this.dilation = builder.dilation;
         this.kernelSize = builder.kernelSize;
         this.stride = builder.stride;
         this.padding = builder.padding;
@@ -55,6 +57,8 @@ public abstract class BaseConvolutionLayerSpace<T extends FeedForwardLayer> exte
 
     protected void setLayerOptionsBuilder(ConvolutionLayer.Builder builder, double[] values) {
         super.setLayerOptionsBuilder(builder, values);
+        if (dilation != null)
+            builder.dilation(dilation.getValue(values));
         if (kernelSize != null)
             builder.kernelSize(kernelSize.getValue(values));
         if (stride != null)
@@ -75,6 +79,8 @@ public abstract class BaseConvolutionLayerSpace<T extends FeedForwardLayer> exte
     @Override
     public String toString(String delim) {
         StringBuilder sb = new StringBuilder();
+        if (dilation != null)
+            sb.append("dilation: ").append(dilation).append(delim);
         if (kernelSize != null)
             sb.append("kernelSize: ").append(kernelSize).append(delim);
         if (stride != null)
@@ -91,12 +97,21 @@ public abstract class BaseConvolutionLayerSpace<T extends FeedForwardLayer> exte
 
 
     public static abstract class Builder<T> extends FeedForwardLayerSpace.Builder<T> {
+        protected ParameterSpace<int[]> dilation;
         protected ParameterSpace<int[]> kernelSize;
         protected ParameterSpace<int[]> stride;
         protected ParameterSpace<int[]> padding;
         protected ParameterSpace<ConvolutionMode> convolutionMode;
         protected ParameterSpace<Boolean> hasBias;
 
+        public Builder dilation(int... dilation) {
+            return dilation(new FixedValue<>(dilation));
+        }
+
+        public Builder dilation(ParameterSpace<int[]> dilation) {
+            this.dilation = dilation;
+            return this;
+        }
         public T kernelSize(int... kernelSize) {
             return kernelSize(new FixedValue<>(kernelSize));
         }
